requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.25.6
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Automat==22.10.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
cryptography==42.0.5
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
ecdsa==0.19.0
executing==2.0.1
fastapi==0.104.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
freezegun==1.4.0
frozendict==2.4.2
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
gradio==4.27.0
gradio_client==0.15.1
grpcio==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
itemadapter==0.8.0
itemloaders==1.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.0
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.16
langchain-anthropic==0.1.11
langchain-community==0.0.34
langchain-core==0.1.45
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.49
llvmlite==0.42.0
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.23.2
openbb==4.1.3
openbb-benzinga==1.1.5
openbb-core==1.1.6
openbb-crypto==1.1.5
openbb-currency==1.1.5
openbb-derivatives==1.1.5
openbb-economy==1.1.5
openbb-equity==1.1.5
openbb-etf==1.1.5
openbb-federal-reserve==1.1.5
openbb-fixedincome==1.1.5
openbb-fmp==1.1.5
openbb-fred==1.1.5
openbb-index==1.1.5
openbb-intrinio==1.1.5
openbb-news==1.1.5
openbb-oecd==1.1.1
openbb-polygon==1.1.5
openbb-regulators==1.1.5
openbb-sec==1.1.5
openbb-tiingo==1.1.5
openbb-tradingeconomics==1.1.5
openbb-yfinance==1.1.5
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
peewee==3.17.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.8.2
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
Protego==0.3.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.5.0
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.17.2
pynndescent==0.5.12
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pytest==8.1.1
pytest-freezegun==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
qdrant-client==1.9.0
queuelib==1.6.2
referencing==0.34.0
regex==2024.4.16
reportlab==4.2.0
requests==2.31.0
requests-cache==1.2.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.4.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
Scrapy==2.11.1
sec-api==1.0.18
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.27.0
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.4.0
tiktoken==0.6.0
tinycss2==1.2.1
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.0
Twisted==24.3.0
typer==0.12.3
types-python-dateutil==2.9.0.20240316
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
umap-learn==0.5.6
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.2.1
uuid7==0.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
yfinance==0.2.38
zipp==3.18.1
zope.interface==6.3
